requirements.txt
tkinter

DateTime
Pillow
pipwin
pyttsx3
pywikihow
pycopy-webbrowser
SpeechRecognition